[REMOVE] config/envValue in ui-mui-react-example
