+++ apps/nextjs-client-template/src/design-system/components/layout/layout-dashboard.tsx
@@ -18,7 +18,6 @@ import Toolbar from '@mui/material/Toolbar';
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
 
-import { websiteName } from '../../config/envValue';
 const drawerWidth = 240;
 
 const openedMixin = (theme: Theme): CSSObject => ({
@@ -131,6 +130,7 @@ export interface LayoutDashboardProps {
     route: string;
     Icon?: typeof MenuIcon;
   }[];
+  websiteName?: string;
 }
 export default function LayoutDashboard({
   children,
@@ -139,6 +139,7 @@ export default function LayoutDashboard({
   logout,
   username,
   menus,
+  websiteName,
 }: LayoutDashboardProps) {
   const theme = useTheme();
 
@@ -170,7 +171,7 @@ export default function LayoutDashboard({
           </IconButton>
           <Logo src={logo} alt="logo-appbar" open={open} />
           <Typography variant="h6" noWrap component="div">
-            {websiteName}
+            {websiteName || null}
           </Typography>
           <Box sx={{ flexGrow: 1 }} />
           {isAuthenticated && (
