--- apps/nextjs-client-template/src/design-system/config/envValue.ts
@@ -1,1 +1,0 @@
-export const websiteName = process.env['NEXT_PUBLIC_WEBSITE_NAME'];
