+++ apps/react-dashboard-template/template/src/app/pages/RootRoutes.tsx
@@ -1,29 +1,11 @@
 import React from 'react';
-import { Routes, Route, Navigate, useLocation } from 'react-router-dom';
+import { Routes, Route } from 'react-router-dom';
 import LayoutDashboard from '../../design-system/components/layout/layout-dashboard-csr';
-import { useAuth } from '../../utils/auth-strapi';
+import { useAuth, useProtectedRoutes } from '../../utils/auth-strapi';
 
+import { websiteName } from '../../config/envValue';
 import logo from '../../assets/img/logo.svg';
 import { menus } from '../config/menus';
-
-const useElementBuilder = (
-  Component: React.LazyExoticComponent<() => JSX.Element>,
-  options?: {
-    isProtected: boolean;
-  }
-) => {
-  const { isAuthenticated } = useAuth();
-  const location = useLocation();
-
-  if (options?.isProtected && !isAuthenticated) {
-    return <Navigate to="/" state={{ from: location }} replace />;
-  }
-  return (
-    <React.Suspense fallback={<>...</>}>
-      <Component />
-    </React.Suspense>
-  );
-};
 
 const Home = React.lazy(() => import('../pages/home'));
 const Common = React.lazy(() => import('../../features/common/pages'));
@@ -41,14 +23,15 @@ const RootRoutes = () => {
             logout={logout}
             isAuthenticated={isAuthenticated}
             username={user?.username}
+            websiteName={websiteName}
           />
         }
       >
-        <Route index element={useElementBuilder(Home)} />
+        <Route index element={useProtectedRoutes(Home)} />
         <Route path="*" element={<>No page</>} />
         <Route
           path="/common"
-          element={useElementBuilder(Common, { isProtected: true })}
+          element={useProtectedRoutes(Common, { isProtected: true })}
         />
       </Route>
     </Routes>
