--- apps/react-dashboard-template/template/src/design-system/config/envValue.ts
@@ -1,1 +1,0 @@
-export const websiteName = process.env['REACT_APP_WEBSITE_NAME'];
