+++ libs/custom-generator/src/generators/react-dashboard-template/files/src/app/pages/RootRoutes.tsx
@@ -2,33 +2,13 @@ import React from 'react'
 import {
   Routes,
   Route,
-  Navigate,
-  useLocation,
 } from "react-router-dom";
 import LayoutDashboard from '../../design-system/components/layout/layout-dashboard-csr';
-import { useAuth } from '../../utils/auth-strapi';
+import { useAuth, useProtectedRoutes } from '../../utils/auth-strapi';
 
+import { websiteName } from '../../config/envValue'
 import logo from '../../assets/img/logo.svg'
 import { menus } from '../config/menus'
-
-const useElementBuilder = (
-  Component: React.LazyExoticComponent<() => JSX.Element>,
-  options?: {
-    isProtected: boolean
-  },
-) => {
-  const { isAuthenticated } = useAuth()
-  const location = useLocation()
-
-  if (options?.isProtected && !isAuthenticated) {
-    return <Navigate to="/" state={{ from: location }} replace />;
-  }
-  return (
-    <React.Suspense fallback={<>...</>}>
-      <Component />
-    </React.Suspense>
-  )
-}
 
 const Home = React.lazy(() => import('../pages/home'))
 const Common = React.lazy(() => import('../../features/common/pages'));
@@ -44,11 +24,12 @@ const RootRoutes = () => {
           logout={logout}
           isAuthenticated={isAuthenticated}
           username={user?.username}
+          websiteName={websiteName}
         />
       )}>
-        <Route index element={useElementBuilder(Home)} />
+        <Route index element={useProtectedRoutes(Home)} />
         <Route path="*" element={<>No page</>} />
-        <Route path="/common" element={useElementBuilder(Common, { isProtected: true })} />
+        <Route path="/common" element={useProtectedRoutes(Common, { isProtected: true })} />
       </Route>
     </Routes>
   )
