+++ libs/custom-generator/src/generators/react-dashboard-template/files/src/utils/auth-strapi/lib/auth-strapi.tsx
@@ -1,7 +1,8 @@
-import { createAuthentication } from '@<%= npmScope %>/util-auth'
+import React from 'react'
+import { Navigate, useLocation, useNavigate } from 'react-router-dom';
 
+import { createAuthentication } from '@<%= npmScope %>/util-auth'
 import apiStrapi, { strapiTokenKey } from '@<%= npmScope %>/data-access-strapi';
-import { useNavigate } from 'react-router-dom';
 
 const { AuthContext, AuthProvider, useAuth } = createAuthentication({
   tokenKey: strapiTokenKey,
@@ -23,8 +24,28 @@ const AuthProviderApp = ({
   )
 }
 
+const useProtectedRoutes = (
+  Component: React.LazyExoticComponent<() => JSX.Element>,
+  options?: {
+    isProtected: boolean
+  },
+) => {
+  const { isAuthenticated } = useAuth()
+  const location = useLocation()
+
+  if (options?.isProtected && !isAuthenticated) {
+    return <Navigate to="/" state={{ from: location }} replace />;
+  }
+  return (
+    <React.Suspense fallback={<>...</>}>
+      <Component />
+    </React.Suspense>
+  )
+}
+
 export {
   AuthContext,
   AuthProviderApp,
   useAuth,
+  useProtectedRoutes,
 }
