+++ libs/custom-generator/src/generators/react-dashboard-template/libs/generate-from-modules.ts
@@ -119,13 +119,6 @@ export default function (tree: Tree, options: NormalizedReactDashboardSchema) {
             `${prefixPath}/src/design-system/mui-theme/index.ts`,
           ]
         },
-        {
-          fromString: 'NX_',
-          toString: 'REACT_APP_',
-          paths: [
-            `${prefixPath}/src/design-system/config/envValue.ts`,
-          ]
-        },
       ] : undefined
     })
     addModules({
