--- libs/custom-generator/src/generators/ui-mui-react-template/files/src/config/envValue.ts
@@ -1,1 +1,0 @@
-export const websiteName = process.env['NX_WEBSITE_NAME']
